app/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
replicate